+++ internal/venice-common/src/main/java/com/linkedin/venice/pubsub/manager/TopicManager.java
@@ -2,10 +2,6 @@ package com.linkedin.venice.pubsub.manager;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
-import com.linkedin.venice.kafka.partitionoffset.InstrumentedPartitionOffsetFetcher;
-import com.linkedin.venice.kafka.partitionoffset.PartitionOffsetFetcher;
-import com.linkedin.venice.kafka.partitionoffset.PartitionOffsetFetcherImpl;
-import com.linkedin.venice.kafka.partitionoffset.PartitionOffsetFetcherStats;
 import com.linkedin.venice.kafka.protocol.KafkaMessageEnvelope;
 import com.linkedin.venice.pubsub.PubSubConstants;
 import com.linkedin.venice.pubsub.PubSubConsumerAdapterFactory;
@@ -23,6 +19,10 @@ import com.linkedin.venice.pubsub.api.exceptions.PubSubClientRetriableException;
 import com.linkedin.venice.pubsub.api.exceptions.PubSubOpTimeoutException;
 import com.linkedin.venice.pubsub.api.exceptions.PubSubTopicDoesNotExistException;
 import com.linkedin.venice.pubsub.api.exceptions.PubSubTopicExistsException;
+import com.linkedin.venice.pubsub.manager.partitionoffset.InstrumentedPartitionOffsetFetcher;
+import com.linkedin.venice.pubsub.manager.partitionoffset.PartitionOffsetFetcher;
+import com.linkedin.venice.pubsub.manager.partitionoffset.PartitionOffsetFetcherImpl;
+import com.linkedin.venice.pubsub.manager.partitionoffset.PartitionOffsetFetcherStats;
 import com.linkedin.venice.serialization.avro.KafkaValueSerializer;
 import com.linkedin.venice.utils.ExceptionUtils;
 import com.linkedin.venice.utils.RetryUtils;
