+++ internal/venice-common/src/main/java/com/linkedin/venice/pubsub/manager/partitionoffset/InstrumentedPartitionOffsetFetcher.java
@@ -1,4 +1,4 @@
-package com.linkedin.venice.kafka.partitionoffset;
+package com.linkedin.venice.pubsub.manager.partitionoffset;
 
 import com.linkedin.venice.pubsub.PubSubTopicPartitionInfo;
 import com.linkedin.venice.pubsub.api.PubSubTopic;
